requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
discord.py==2.1.0
disnake==2.7.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
multidict==6.0.3
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
urllib3==1.26.13
yarl==1.8.2
youtube-dl==2021.12.17
youtube-search==2.1.2
